+++ orchestrator.h
@@ -11,5 +11,5 @@ public:
     Orchestrator(OutputStream &stream);
     int start(std::string &address_and_mask);
     void stop();
-    uint16_t get_progress();     // returns the % of finalized hosts
+    int get_progress();     // returns the % of finalized hosts
 };
